Core/DBLayer: multiple fixes
*Fixed .list item command crash
*Fixed Mail system issues
*Fixed petitions crashes
*Fixed faction/race change
*Fixed WP loading query

+++ src/server/game/Globals/ObjectMgr.cpp
@@ -5600,10 +5600,10 @@ void ObjectMgr::ReturnOrDeleteOldMails(bool serverUp)
         m->sender = fields[2].GetUInt32();
         m->receiver = fields[3].GetUInt32();
         bool has_items = fields[4].GetBool();
-        m->expire_time = time_t(fields[5].GetUInt64());
+        m->expire_time = time_t(fields[5].GetUInt32());
         m->deliver_time = 0;
-        m->COD = fields[6].GetUInt64();
-        m->checked = fields[7].GetUInt32();
+        m->COD = fields[6].GetUInt32();
+        m->checked = fields[7].GetUInt8();
         m->mailTemplateId = fields[8].GetInt16();
 
         Player *player = NULL;
